+++ com.unity.cinemachine/Samples~/Shared Assets/Scripts/SimplePlayerController.cs
@@ -7,9 +7,13 @@ namespace Unity.Cinemachine.Samples
 {
     public abstract class SimplePlayerControllerBase : MonoBehaviour, IInputAxisOwner
     {
+        [Tooltip("Ground speed when walking")]
         public float Speed = 1f;
+        [Tooltip("Ground speed when sprinting")]
         public float SprintSpeed = 4;
+        [Tooltip("Initial vertical speed when jumping")]
         public float JumpSpeed = 4;
+        [Tooltip("Initial vertical speed when sprint-jumping")]
         public float SprintJumpSpeed = 6;
 
         public Action PreUpdate;
@@ -45,21 +49,45 @@ namespace Unity.Cinemachine.Samples
             axes.Add(new () { DrivenAxis = () => ref Jump, Name = "Jump" });
             axes.Add(new () { DrivenAxis = () => ref Sprint, Name = "Sprint" });
         }
+
+        public virtual void SetStrafeMode(bool b) {}
+        public abstract bool IsMoving { get; }
     }
 
     public class SimplePlayerController : SimplePlayerControllerBase
     {
+        [Tooltip("How long it takes for the player to change velocity")]
         public float Damping = 0.5f;
+
+        [Tooltip("If true, player will strafe when moving sideways, otherwise will turn to face direction of motion")]
         public bool Strafe = false;
 
         public enum ForwardModes { Camera, Player, World };
-        public ForwardModes InputForward = ForwardModes.Camera;
         public enum UpModes { Player, World };
+
+        [Tooltip("Reference frame for the input controls:\n"
+            + "<b>Camera</b>: Input forward is camera forward direction.\n"
+            + "<b>Player</b>: Input forward is Player's forward direction.\n"
+            + "<b>World</b>: Input forward is World forward direction.")]
+        public ForwardModes InputForward = ForwardModes.Camera;
+
+        [Tooltip("Up direction for computing motion:\n"
+            + "<b>Player</b>: Will move in the Player's local XZ plane.\n"
+            + "<b>World</b>: will move in global XZ plane.")]
         public UpModes UpMode = UpModes.World;
 
         [Tooltip("Override the main camera. Useful for split screen games.")]
         public Camera CameraOverride;
+
+        [Tooltip("Raycasts for ground will detect these layers")]
+        public LayerMask GroundLayers = 1;
         
+        [Tooltip("Force of gravity in the down direction (m/s/s)")]
+        public float Gravity = 10;
+
+        const float kDelayBeforeInferringJump = 0.3f;
+        float m_TimeLastGrounded = 0;
+
         Vector3 m_CurrentVelocityXZ;
         Vector3 m_LastInput;
         float m_CurrentVelocityY;
@@ -67,10 +95,28 @@ namespace Unity.Cinemachine.Samples
         bool m_IsJumping;
         CharacterController m_Controller; // optional
 
+        // These are part of a strategy to combat input gimbal lock when controlling a player
+        // that can move freely on surfaces that go upside-down relative to the camera.
+        bool m_InTopHemisphere = true;
+        float m_TimeInHemisphere = 100;
+        Vector3 m_LastRawInput;
+        Quaternion m_Upsidedown = Quaternion.AngleAxis(180, Vector3.left);
+
+        public override void SetStrafeMode(bool b) => Strafe = b;
+        public override bool IsMoving => m_LastInput.sqrMagnitude > 0.01f;
+
         public bool IsSprinting => m_IsSprinting;
         public bool IsJumping => m_IsJumping;
-        public bool IsMoving => m_LastInput.sqrMagnitude > 0.01f;
         public Camera Camera => CameraOverride == null ? Camera.main : CameraOverride;
+
+        public bool IsGrounded()
+        {
+            if (m_Controller != null)
+                return m_Controller.isGrounded;
+
+            // No controller - must compute manually with raycast
+            return GetDistanceFromGround(transform.position, UpDirection, 10) < 0.01f;
+        }
 
         void Start() => TryGetComponent(out m_Controller);
 
@@ -79,6 +125,7 @@ namespace Unity.Cinemachine.Samples
             m_CurrentVelocityY = 0;
             m_IsSprinting = false;
             m_IsJumping = false;
+            m_TimeLastGrounded = Time.time;
         }
 
         void Update()
@@ -89,11 +136,12 @@ namespace Unity.Cinemachine.Samples
             bool justLanded = ProcessJump();
 
             // Get the reference frame for the input
-            if (!GetInputFrame(out var inputFrame))
-                return; // no input frame, give up
+            var rawInput = new Vector3(MoveX.Value, 0, MoveZ.Value);
+            var inputFrame = GetInputFrame(Vector3.Dot(rawInput, m_LastRawInput) < 0.8f);
+            m_LastRawInput = rawInput;
 
             // Read the input from the user and put it in the input frame
-            m_LastInput = inputFrame * new Vector3(MoveX.Value, 0, MoveZ.Value);
+            m_LastInput = inputFrame * rawInput;
             if (m_LastInput.sqrMagnitude > 1)
                 m_LastInput.Normalize();
 
@@ -118,7 +166,7 @@ namespace Unity.Cinemachine.Samples
             // If not strafing, rotate the player to face movement direction
             if (!Strafe && m_CurrentVelocityXZ.sqrMagnitude > 0.001f)
             {
-                var fwd = inputFrame * new Vector3(0, 0, 1);
+                var fwd = inputFrame * Vector3.forward;
                 var qA = transform.rotation;
                 var qB = Quaternion.LookRotation(
                     (InputForward == ForwardModes.Player && Vector3.Dot(fwd, m_CurrentVelocityXZ) < 0)
@@ -138,40 +186,107 @@ namespace Unity.Cinemachine.Samples
 
         Vector3 UpDirection => UpMode == UpModes.World ? Vector3.up : transform.up;
 
-        // Get the reference frame for the input
-        bool GetInputFrame(out Quaternion inputFrame)
+        // Get the reference frame for the input.  The idea is to map camera fwd/right
+        // to the player's XZ plane.  There is some complexity here to avoid
+        // gimbal lock when the player is tilted 180 degrees relative to the camera.
+        Quaternion GetInputFrame(bool inputDirectionChanged)
         {
-            var up = UpDirection;
-            var fwd = InputForward switch
+            // Get the raw input frame, depending of forward mode setting
+            var frame = Quaternion.identity;
+            switch (InputForward)
             {
-                ForwardModes.Camera => Camera.transform.forward,
-                ForwardModes.Player => transform.forward,
-                _ => Vector3.forward,
-            };
-            fwd = fwd.ProjectOntoPlane(up);
-            if (fwd.sqrMagnitude < 0.001f)
-            {
-                inputFrame = Quaternion.identity;
-                return false;
+                case ForwardModes.Camera: frame = Camera.transform.rotation; break;
+                case ForwardModes.Player: return transform.rotation;
+                case ForwardModes.World: break;
             }
-            inputFrame = Quaternion.LookRotation(fwd, up);
-            return true;
+
+            // Map the raw input frame to something that makes sense as a direction for the player
+            var playerUp = transform.up;
+            var up = frame * Vector3.up;
+
+            // Is the player in the top or bottom hemisphere?  This is needed to avoid gimbal lock
+            const float BlendTime = 2f;
+            m_TimeInHemisphere += Time.deltaTime;
+            bool inTopHemisphere = Vector3.Dot(up, playerUp) >= 0;
+            if (inTopHemisphere != m_InTopHemisphere)
+            {
+                m_InTopHemisphere = inTopHemisphere;
+                m_TimeInHemisphere = Mathf.Max(0, BlendTime - m_TimeInHemisphere);
+            }
+
+            // If the player is untilted relative to the input frmae, then early-out with a simple LookRotation
+            var axis = Vector3.Cross(up, playerUp);
+            if (axis.sqrMagnitude < 0.001f && inTopHemisphere)
+                return frame;
+
+            // Player is tilted relative to input frame: tilt the input frame to match
+            var angle = UnityVectorExtensions.SignedAngle(up, playerUp, axis);
+            var frameA = Quaternion.AngleAxis(angle, axis) * frame;
+
+            // If the player is tilted, then we need to get tricky to avoid gimbal-lock
+            // when player is tilted 180 degrees.  There is no perfect solution for this,
+            // we need to cheat it :/
+            Quaternion frameB = frameA;
+            if (!inTopHemisphere || m_TimeInHemisphere < BlendTime)
+            {
+                // Compute an alternative reference frame for the bottom hemisphere.
+                // The two reference frames are incompatible where they meet, especially
+                // when player up is pointing along the X axis of camera frame. 
+                // There is no one reference frame that works for all player directions.
+                frameB = frame * m_Upsidedown;
+                var axisB = Vector3.Cross(frameB * Vector3.up, playerUp);
+                if (axisB.sqrMagnitude > 0.001f)
+                    frameB = Quaternion.AngleAxis(180f - angle, axisB) * frameB;
+            }
+            // Blend timer force-expires when user changes input direction
+            if (inputDirectionChanged)
+                m_TimeInHemisphere = BlendTime;
+
+            // If we have been long enough in one hemisphere, then we can just use its reference frame
+            if (m_TimeInHemisphere >= BlendTime)
+                return inTopHemisphere ? frameA : frameB;
+
+            // Because frameA and frameB do not join seamlessly when player Up is along X axis,
+            // we blend them over a time in order to avoid degenerate spinning.
+            // This will produce weird movements occasionally, but it's the lesser of the evils.
+            if (inTopHemisphere)
+                return Quaternion.Slerp(frameB, frameA, m_TimeInHemisphere / BlendTime);
+            return Quaternion.Slerp(frameA, frameB, m_TimeInHemisphere / BlendTime);
         }
 
         bool ProcessJump()
         {
-            const float kGravity = -10;
             bool justLanded = false;
-            m_CurrentVelocityY += kGravity * Time.deltaTime;
-            if (!m_IsJumping && Jump.Value > 0.01f)
+            var now = Time.time;
+            bool grounded = IsGrounded();
+
+            m_CurrentVelocityY -= Gravity * Time.deltaTime;
+
+            if (!m_IsJumping)
             {
-                m_IsJumping = true;
-                StartJump?.Invoke();
-                m_CurrentVelocityY = m_IsSprinting ? SprintJumpSpeed : JumpSpeed;
+                // Process jump command
+                if (grounded && Jump.Value > 0.01f)
+                {
+                    m_IsJumping = true;
+                    m_CurrentVelocityY = m_IsSprinting ? SprintJumpSpeed : JumpSpeed;
+                }
+                // If we are falling, assume the jump pose
+                if (!grounded && now - m_TimeLastGrounded > kDelayBeforeInferringJump)
+                    m_IsJumping = true;
+ 
+                if (m_IsJumping)
+                {
+                    StartJump?.Invoke();
+                    grounded = false;
+                }
             }
-            if (IsGrounded() && m_CurrentVelocityY < 0)
+
+            if (grounded)
             {
+                m_TimeLastGrounded = Time.time;
                 m_CurrentVelocityY = 0;
+
+                // If we were jumping, complete the jump
                 if (m_IsJumping)
                 {
                     EndJump?.Invoke();
@@ -181,15 +296,6 @@ namespace Unity.Cinemachine.Samples
                 }
             }
             return justLanded;
-        }
-
-        bool IsGrounded()
-        {
-            if (m_Controller != null)
-                return m_Controller.isGrounded;
-
-            // No controller - must compute manually with raycast
-            return GetDistanceFromGround(transform.position, UpDirection, 10) < 0.01f;
         }
 
         void ApplyMotion()
@@ -224,7 +330,7 @@ namespace Unity.Cinemachine.Samples
         float GetDistanceFromGround(Vector3 pos, Vector3 up, float max)
         {
             float kExtraHeight = 2;
-            if (Physics.Raycast(pos + up * kExtraHeight, -up, out var hit, max + kExtraHeight, LayerMask.GetMask("Default")))
+            if (Physics.Raycast(pos + up * kExtraHeight, -up, out var hit, max + kExtraHeight, GroundLayers))
                 return hit.distance - kExtraHeight; 
             return max + 1;
         }