+++ com.unity.cinemachine/CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - ThirdPersonFollow sets state.ReferenceUp, to allow usage on arbitrary surfaces without global world up override.
-
+- SimplePlayerAimController sample upgraded to work on arbitrary surfaces (no longer depends on world up).
+- SimplePlayerController sample has logic to avoid input gimbal lock when player is upside-down relative to the camera.
+- PlayerOnSphere sample is now PlayerOnSurface - can walk on arbitrary surfaces.
 
 ## [3.0.1] - 2023-11-27